refactor(letter-grid): extract shared grid/path/camera engine into _grid.js
Pulls v1's virtual-grid hash, alphabet resolution, word-path search and the
reading-head camera math out of letter-grid-v1-spell/index.js into a pure,
p5-free module shared by the whole letter-grid category. v1 now imports it and
renders only the flat 2D view; behaviour is unchanged (same path, seamless loop,
verified against the pre-refactor output). Prepares the ground for the extruded
v2 variant which reuses the same engine.
